+++ starlight-core/src/main/java/org/atcraftmc/starlight/core/view/PlayerViewChannelRenderer.java
@@ -1,38 +1,79 @@
 package org.atcraftmc.starlight.core.view;
 
-import org.atcraftmc.qlib.bukkit.task.Task;
-import org.atcraftmc.qlib.bukkit.task.TaskScheduler;
 import org.atcraftmc.starlight.core.PlayerView;
-import org.atcraftmc.starlight.core.TaskService;
+import org.atcraftmc.starlight.core.view.process.TaskScheduleProcess;
+import org.atcraftmc.starlight.core.view.process.ViewRenderProcess;
+import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Consumer;
 
 public class PlayerViewChannelRenderer {
+    private final Map<String, ViewRenderProcess> renderers = new HashMap<>();
+    private final AtomicReference<ViewRenderProcess> current = new AtomicReference<>();
     private final PlayerView holder;
-    private final Map<String, PlayerViewRenderer> renderers = new HashMap<>();
+    private final Set<String> rejects = new HashSet<>();
+    private final UUID player;
+    private boolean rejectAll = false;
+    private Consumer<Player> cleanupAction = (p) -> {};
 
-    private boolean rejectAll;
-    private Task currentTask;
 
     public PlayerViewChannelRenderer(PlayerView holder) {
         this.holder = holder;
+        this.player = holder.pointer().getUniqueId();//todo
     }
 
-    private void select() {
-        if (this.currentTask != null) {
-            this.currentTask.cancel();
+    public void setCleanupAction(Consumer<Player> cleanupAction) {
+        this.cleanupAction = cleanupAction;
+    }
+
+    public Optional<ViewRenderProcess> getCurrent() {
+        return Optional.ofNullable(this.current.get());
+    }
+
+    public ViewRenderProcess registerProcess(String id, ViewRenderProcess process) {
+        this.renderers.put(id, process);
+        this.update();
+
+        return process;
+    }
+
+    public ViewRenderProcess registerIntervalProcess(String id, int priority, int interval, SchedulerProvider provider, ViewRendererCallback func) {
+        return this.registerProcess(id, new TaskScheduleProcess(this.player, id, priority, interval, provider, func));
+    }
+
+    public void addReject(String id) {
+        this.rejects.add(id);
+    }
+
+    public void removeReject(String id) {
+        this.rejects.remove(id);
+    }
+
+
+    private void update() {
+        var previous = this.getCurrent();
+        var player = Bukkit.getPlayer(this.player);
+
+        if (previous.isPresent()) {
+            previous.get().inactive(player);
+            this.current.set(null);
         }
 
+
         if (this.rejectAll) {
+            this.cleanupAction.accept(player);
             return;
         }
 
         var list = new ArrayList<>(this.renderers.values());
 
+        list.removeIf((p) -> this.rejects.contains(p.id()));
+
         if (list.isEmpty()) {
+            this.cleanupAction.accept(player);
             return;
         }
 
@@ -41,7 +82,7 @@ public class PlayerViewChannelRenderer {
                 return 0;
             }
 
-            int pri = -Comparator.comparingInt(PlayerViewRenderer::priority).compare(o1, o2);
+            int pri = -Comparator.comparingInt(ViewRenderProcess::priority).compare(o1, o2);
 
             if (pri != 0) {
                 return pri;
@@ -52,17 +93,9 @@ public class PlayerViewChannelRenderer {
 
         var selected = list.get(0);
 
-        /*
+        this.current.set(selected);
 
-        this.currentTask = selected.scheduler().apply(this.holder.pointer()).timer(1, selected.interval(), (t) -> {
-            if (this.holder.isChannelRejected(selected.id())) {
-                return;
-            }
-
-            selected.renderer().render(this.holder.pointer(), t);
-        });
-
-         */
+        selected.active(player);
     }
 
     public void rejectAll(boolean enable) {